+++ frontend/src/app/admin/page.tsx
@@ -3,7 +3,7 @@
 import { useEffect } from 'react';
 import { useRouter } from 'next/navigation';
 import Link from 'next/link';
-import { useAuth } from '../context/AuthContext';
+import { useAuth } from '../../context/AuthContext';
 
 export default function AdminPage() {
   const { isAuthenticated, isAdmin } = useAuth();
