+++ frontend/src/services/productService.ts
